feat: create fetch wrapper class
This fetch wrapper is born more from an experiment to play with method-chaining than from a real need.

However the wrapper helps to reduce boilerplate code for handling errors and exceptions and moreover provides a clean and straightforward syntax which helps speeding up coding.

I will therefore use it wherever it makes sense to.