chore: install `swr` library for client-side data fetching
`swr` is developed by Vercel, like Next.js, and is a recommended library to easily perform data fetching on the client-side.
Before using this library, I tried the `use()` hook from React, but I didn't manage to make it get along well with Next.js.
Moreover, `swr` has interesting functionalities like auto-refreshing data when the user comes back on the tab and more, so I'll give it a go.