refactor: avoid `<p>` nested in another `<p>` tag per Next.js docs
Next.js can, in certain circumstances, throw a hydration error if it tries to hydrate the page but encounters an incorrect nesting of HTML tags.

I got the error, but surprisingly only on iPhone and only on the `object-detection` page.

More on this error in the Next.js docs: https://nextjs.org/docs/messages/react-hydration-error